feat: migrate to web app
