+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "lipoic-frontend",
+  "name": "lipoic-web-app",
   "private": true,
   "version": "0.0.0",
   "scripts": {
