+++ src/components/MainView.vue
@@ -1,32 +1,7 @@
-<script lang="ts" setup>
-import { defineAsyncComponent, ref } from 'vue';
-
-const BlockHeaderVue = defineAsyncComponent(
-  () => import('@/components/header/BlockHeader.vue')
-);
-const BlockFooterVue = defineAsyncComponent(
-  () => import('@/components/BlockFooter.vue')
-);
-// Child event
-const headerRef = ref<InstanceType<typeof BlockHeaderVue> | null>(null);
-const headerMenuHandler = (): void => {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  headerRef.value?.changeMenuCheckboxState();
-};
-</script>
+<script lang="ts" setup></script>
 
 <template>
-  <div class="main">
-    <header>
-      <BlockHeaderVue ref="headerRef" />
-    </header>
-    <main class="wrapper" @touchstart.passive="headerMenuHandler">
-      <slot />
-    </main>
-    <footer>
-      <BlockFooterVue />
-    </footer>
-  </div>
+  <div class="main"></div>
 </template>
 
 <style lang="scss" scoped>
@@ -37,17 +12,5 @@ const headerMenuHandler = (): void => {
   flex-direction: column;
   min-height: 100%;
   background-color: $LightBlack;
-
-  .wrapper {
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-    flex-grow: 1;
-    justify-content: flex-start;
-
-    :deep() > * {
-      height: 100%;
-    }
-  }
 }
 </style>
