+++ src/locales/zh-CN.json
@@ -36,31 +36,5 @@
   "error": {
     "message": "发生未知错误，请稍后再试。"
   },
-  "footer": {
-    "copyright": "版权所有 © 2022 Lipoic. 保留所有权利。"
-  },
-  "header": {
-    "action": {
-      "logout": "注销"
-    },
-    "links": {
-      "about": "关于",
-      "demo": "演示",
-      "feature": "特色",
-      "home": "首页"
-    },
-    "login": "登錄"
-  },
-  "home": {
-    "login": {
-      "parents": "家长",
-      "student": "学生",
-      "teacher": "老师"
-    },
-    "subtitle": {
-      "1": "专门为师生们设计的多功能教育平台",
-      "2": "熱愛開放原始碼與開放教育精神"
-    }
-  },
   "name": "简体中文 (中国)"
 }
