+++ src/locales/zh-TW.json
@@ -10,7 +10,7 @@
         "user": "此使用者不存在，請檢查您的電子郵件地址是否正確輸入。",
         "verifyEmail": "此電子郵件尚未驗證，請到您的電子郵件信箱檢查信件。"
       },
-      "forgotPassword": "忘記密碼",
+      "forgotPassword": "忘記密碼？",
       "haveNoAccount": "還沒有帳號嗎？",
       "loading": "登入中...",
       "loginButton": "登入",
@@ -35,32 +35,6 @@
   },
   "error": {
     "message": "發生未知錯誤，請稍後再試。"
-  },
-  "footer": {
-    "copyright": "版權所有 © 2022 Lipoic. 保留所有權利。"
-  },
-  "header": {
-    "action": {
-      "logout": "登出"
-    },
-    "links": {
-      "about": "關於",
-      "demo": "展示",
-      "feature": "特色",
-      "home": "首頁"
-    },
-    "login": "登入"
-  },
-  "home": {
-    "login": {
-      "parents": "家長",
-      "student": "學生",
-      "teacher": "老師"
-    },
-    "subtitle": {
-      "1": "專門為師生們設計的多功能教育平台",
-      "2": "熱愛開放原始碼與開放教育精神"
-    }
   },
   "name": "繁體中文 (台灣)"
 }
