+++ src/views/PageOverview.vue
@@ -1,17 +1,22 @@
 <script lang="ts" setup>
+import { defineAsyncComponent } from 'vue';
 import { useRouter } from 'vue-router';
 import { useUserStore } from '@/stores/models/user';
 
-// If the user is logged in, redirect to the overview page
+const MainViewVue = defineAsyncComponent(
+  () => import('@/components/MainView.vue')
+);
+
 const isLogin = useUserStore().isLoggedIn();
 const router = useRouter();
-if (isLogin) {
-  router.push('/overview');
-} else {
+
+if (!isLogin) {
   router.push('/account/login');
 }
 </script>
 
 <template>
-  <div></div>
+  <MainViewVue />
 </template>
+
+<style lang="scss" scoped></style>
